# chat-api/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Chat Service
  description: Service for chat app, a Goa teaser
  version: ""
host: 172.26.0.3:8000
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
  /chatroom/chat:
    post:
      tags:
      - chatapi
      summary: postchat chatapi
      operationId: chatapi#postchat
      parameters:
      - name: PostchatRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/ChatapiPostchatRequestBody'
          required:
          - Id
          - UserId
          - RoomName
          - Member
          - Chat
          - PostDt
          - Cookie
      responses:
        "200":
          description: OK response.
          schema:
            type: boolean
      schemes:
      - http
definitions:
  ChatapiPostchatRequestBody:
    title: ChatapiPostchatRequestBody
    type: object
    properties:
      Chat:
        type: string
        description: chat
        example: Odio recusandae atque ut.
      Cookie:
        type: string
        description: cookie
        example: Error inventore quia molestias dolorem.
      Id:
        type: string
        description: room id
        example: Expedita recusandae vel.
      Member:
        type: string
        description: member
        example: Reprehenderit neque nostrum et magni eos enim.
      PostDt:
        type: string
        example: "1995-09-10T11:00:58Z"
        format: date-time
      RoomName:
        type: string
        description: room name
        example: Est ullam quis repudiandae perferendis.
      UserId:
        type: string
        description: user id
        example: Perspiciatis numquam nostrum voluptas perferendis et.
    example:
      Chat: Sit iusto vero suscipit voluptatibus rerum qui.
      Cookie: Necessitatibus soluta.
      Id: Quo itaque ipsa reiciendis delectus iusto.
      Member: Laudantium vero.
      PostDt: "2015-11-07T07:31:20Z"
      RoomName: Quidem labore.
      UserId: Aut accusamus aut voluptas dicta totam.
    required:
    - Id
    - UserId
    - RoomName
    - Member
    - Chat
    - PostDt
    - Cookie